Route shared host code and the QMixer trace clock through the platform layer; audit long for LLP64
long audit: ftell results (fixture.cpp, loader.cpp) kept, they are what ftell
returns; strtol/strtoul parsing in boot.cpp and script.cpp kept, values are
small; color distance in settings_page.cpp kept, fits 32 bits;
display_settings.cpp mode numbers kept; x86.h lround kept, the quotient is
already reduced to 0..7; vector iterator casts in dx and mods changed from
(long) to (ptrdiff_t).

+++ src/recomp/dx/d3d.cpp
@@ -1146,7 +1146,7 @@ void Viewport_DeleteLight(X86 *c) {
     }
     for (size_t i = 0; i < v->lights.size(); ++i) {
         if (v->lights[i] == l->id) {
-            v->lights.erase(v->lights.begin() + (long)i);
+            v->lights.erase(v->lights.begin() + (ptrdiff_t)i);
             com_release(l);
             com_ret(c, D3D_OK_);
             return;
@@ -1309,7 +1309,7 @@ void Device_DeleteViewport(X86 *c) {
     }
     for (size_t i = 0; i < dev->viewports.size(); ++i) {
         if (dev->viewports[i] == v->id) {
-            dev->viewports.erase(dev->viewports.begin() + (long)i);
+            dev->viewports.erase(dev->viewports.begin() + (ptrdiff_t)i);
             if (dev->current_viewport == v->id)
                 dev->current_viewport = 0;
             com_release(v);
